--- packages/syft/src/syft/core/node/new/task/util.py
@@ -1,33 +1,0 @@
-"""Utility functions"""
-
-# stdlib
-import socket
-
-
-def find_available_port(host: str, port: int, search: bool = False) -> int:
-    port_available = False
-    while not port_available:
-        try:
-            sock = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
-            result_of_check = sock.connect_ex((host, port))
-
-            if result_of_check != 0:
-                port_available = True
-                break
-            else:
-                if search:
-                    port += 1
-                else:
-                    break
-
-        except Exception as e:
-            print(f"Failed to check port {port}. {e}")
-    sock.close()
-
-    if search is False and port_available is False:
-        error = (
-            f"{port} is in use, either free the port or "
-            + f"try: {port}+ to auto search for a port"
-        )
-        raise Exception(error)
-    return port
